+++ include/alpaka/block/shared/dyn/Traits.hpp
@@ -27,20 +27,26 @@ namespace alpaka
         //#############################################################################
         //! The block shared dynamic memory get trait.
         template<typename T, typename TBlockSharedMemDyn, typename TSfinae = void>
-        struct GetMem;
+        struct GetDynSharedMem;
     } // namespace traits
 
     //-----------------------------------------------------------------------------
-    //! Returns the pointr to the block shared dynamic memory.
+    //! Get block shared dynamic memory.
+    //!
+    //! The available size of the memory can be defined by specializing the trait
+    //! BlockSharedMemDynSizeBytes for a kernel.
+    //! The Memory can be accessed by all threads within a block.
+    //! Access to the memory is not thread safe.
     //!
     //! \tparam T The element type.
     //! \tparam TBlockSharedMemDyn The block shared dynamic memory implementation type.
     //! \param blockSharedMemDyn The block shared dynamic memory implementation.
+    //! \return Pointer to pre-allocated contiguous memory.
     ALPAKA_NO_HOST_ACC_WARNING
     template<typename T, typename TBlockSharedMemDyn>
-    ALPAKA_FN_ACC auto getMem(TBlockSharedMemDyn const& blockSharedMemDyn) -> T*
+    ALPAKA_FN_ACC auto getDynSharedMem(TBlockSharedMemDyn const& blockSharedMemDyn) -> T*
     {
         using ImplementationBase = concepts::ImplementationBase<ConceptBlockSharedDyn, TBlockSharedMemDyn>;
-        return traits::GetMem<T, ImplementationBase>::getMem(blockSharedMemDyn);
+        return traits::GetDynSharedMem<T, ImplementationBase>::getMem(blockSharedMemDyn);
     }
 } // namespace alpaka