+++ include/alpaka/block/shared/st/BlockSharedMemStNoSync.hpp
@@ -52,10 +52,10 @@ namespace alpaka
 #endif
         //#############################################################################
         template<typename T, std::size_t TuniqueId>
-        struct AllocVar<T, TuniqueId, BlockSharedMemStNoSync>
+        struct DeclareSharedVar<T, TuniqueId, BlockSharedMemStNoSync>
         {
             //-----------------------------------------------------------------------------
-            ALPAKA_FN_HOST static auto allocVar(BlockSharedMemStNoSync const& blockSharedMemSt) -> T&
+            ALPAKA_FN_HOST static auto declareVar(BlockSharedMemStNoSync const& blockSharedMemSt) -> T&
             {
                 constexpr std::size_t alignmentInBytes = std::max(core::vectorization::defaultAlignment, alignof(T));
 
@@ -69,10 +69,10 @@ namespace alpaka
 #endif
         //#############################################################################
         template<>
-        struct FreeMem<BlockSharedMemStNoSync>
+        struct FreeSharedVars<BlockSharedMemStNoSync>
         {
             //-----------------------------------------------------------------------------
-            ALPAKA_FN_HOST static auto freeMem(BlockSharedMemStNoSync const& blockSharedMemSt) -> void
+            ALPAKA_FN_HOST static auto freeVars(BlockSharedMemStNoSync const& blockSharedMemSt) -> void
             {
                 blockSharedMemSt.m_sharedAllocs.clear();
             }