+++ include/alpaka/kernel/TaskKernelCpuFibers.hpp
@@ -190,7 +190,7 @@ namespace alpaka
             acc.m_fibersToIndices.clear();
 
             // After a block has been processed, the shared memory has to be deleted.
-            freeMem(acc);
+            freeSharedVars(acc);
         }
         //-----------------------------------------------------------------------------
         //! The function executed for each block thread.