+++ include/alpaka/kernel/TaskKernelCpuSerial.hpp
@@ -114,7 +114,7 @@ namespace alpaka
                 boundKernelFnObj(acc);
 
                 // After a block has been processed, the shared memory has to be deleted.
-                freeMem(acc);
+                freeSharedVars(acc);
             });
         }