+++ include/alpaka/kernel/TaskKernelCpuTbbBlocks.hpp
@@ -120,7 +120,7 @@ namespace alpaka
 
                 boundKernelFnObj(acc);
 
-                freeMem(acc);
+                freeSharedVars(acc);
             });
         }