+++ include/alpaka/kernel/TaskKernelCpuThreads.hpp
@@ -181,7 +181,7 @@ namespace alpaka
             acc.m_threadToIndexMap.clear();
 
             // After a block has been processed, the shared memory has to be deleted.
-            freeMem(acc);
+            freeSharedVars(acc);
         }
         //-----------------------------------------------------------------------------
         //! The function executed for each block thread on the host.